5일차 ex01 int-> double로 변경

+++ src/main/java/org/example/Head04_JavaProgrammingStart/example01/PaymentProcessorTest.java
@@ -8,7 +8,7 @@ public class PaymentProcessorTest {
                 new MobilePayment()
         };
 
-        int[] amounts = {
+        double[] amounts = {
                 30000,
                 10000,
                 20000
